+++ nes-core/tests/UnitTests/Optimizer/Phases/TypeInferencePhaseTest.cpp
@@ -40,6 +40,7 @@
 #include <Optimizer/Phases/TypeInferencePhaseContext.hpp>
 #include <Plans/Query/QueryPlan.hpp>
 #include <Topology/TopologyNode.hpp>
+#include <Util/JavaUdfDescriptorBuilder.hpp>
 #include <Util/Logger/Logger.hpp>
 #include <Windowing/TimeCharacteristic.hpp>
 #include <Windowing/WindowActions/LazyNestLoopJoinTriggerActionDescriptor.hpp>
@@ -1486,12 +1487,10 @@ TEST_F(TypeInferencePhaseTest, inferTypeForQueryWithMapUDF) {
 
     auto sinkOperator = LogicalOperatorFactory::createSinkOperator(NullOutputSinkDescriptor::create());
 
-    auto javaUdfDescriptor = std::make_shared<Catalogs::UDF::JavaUdfDescriptor>(
-        "some_class",
-        "some_method",
-        Catalogs::UDF::JavaSerializedInstance{1},
-        Catalogs::UDF::JavaUdfByteCodeList{{"some_class", {1}}},
-        std::make_shared<Schema>()->addField("outputAttribute", DataTypeFactory::createBoolean()));
+    auto javaUdfDescriptor =
+        Catalogs::UDF::JavaUdfDescriptorBuilder{}
+            .setOutputSchema(std::make_shared<Schema>()->addField("outputAttribute", DataTypeFactory::createBoolean()))
+            .build();
     auto mapUdfLogicalOperatorNode =
         std::make_shared<MapJavaUdfLogicalOperatorNode>(javaUdfDescriptor, Util::getNextOperatorId());
 
@@ -1529,12 +1528,10 @@ TEST_F(TypeInferencePhaseTest, inferTypeForQueryWithMapUDFAfterBinaryOperator) {
 
     auto sinkOperator = LogicalOperatorFactory::createSinkOperator(NullOutputSinkDescriptor::create());
 
-    auto javaUdfDescriptor = std::make_shared<Catalogs::UDF::JavaUdfDescriptor>(
-        "some_class",
-        "some_method",
-        Catalogs::UDF::JavaSerializedInstance{1},
-        Catalogs::UDF::JavaUdfByteCodeList{{"some_class", {1}}},
-        std::make_shared<Schema>()->addField("outputAttribute", DataTypeFactory::createBoolean()));
+    auto javaUdfDescriptor =
+        Catalogs::UDF::JavaUdfDescriptorBuilder{}
+            .setOutputSchema(std::make_shared<Schema>()->addField("outputAttribute", DataTypeFactory::createBoolean()))
+            .build();
     auto mapUdfLogicalOperatorNode =
         std::make_shared<MapJavaUdfLogicalOperatorNode>(javaUdfDescriptor, Util::getNextOperatorId());
 
@@ -1580,21 +1577,17 @@ TEST_F(TypeInferencePhaseTest, inferTypeForQueryWithMapUDFBeforeBinaryOperator) 
 
     auto sinkOperator = LogicalOperatorFactory::createSinkOperator(NullOutputSinkDescriptor::create());
 
-    auto javaUdfDescriptor1 = std::make_shared<Catalogs::UDF::JavaUdfDescriptor>(
-        "some_class",
-        "some_method",
-        Catalogs::UDF::JavaSerializedInstance{1},
-        Catalogs::UDF::JavaUdfByteCodeList{{"some_class", {1}}},
-        std::make_shared<Schema>()->addField("outputAttribute1", DataTypeFactory::createBoolean()));
+    auto javaUdfDescriptor1 =
+        Catalogs::UDF::JavaUdfDescriptorBuilder{}
+            .setOutputSchema(std::make_shared<Schema>()->addField("outputAttribute1", DataTypeFactory::createBoolean()))
+            .build();
     auto mapUdfLogicalOperatorNode1 =
         std::make_shared<MapJavaUdfLogicalOperatorNode>(javaUdfDescriptor1, Util::getNextOperatorId());
 
-    auto javaUdfDescriptor2 = std::make_shared<Catalogs::UDF::JavaUdfDescriptor>(
-        "some_class",
-        "some_method",
-        Catalogs::UDF::JavaSerializedInstance{1},
-        Catalogs::UDF::JavaUdfByteCodeList{{"some_class", {1}}},
-        std::make_shared<Schema>()->addField("outputAttribute2", DataTypeFactory::createBoolean()));
+    auto javaUdfDescriptor2 =
+        Catalogs::UDF::JavaUdfDescriptorBuilder{}
+            .setOutputSchema(std::make_shared<Schema>()->addField("outputAttribute2", DataTypeFactory::createBoolean()))
+            .build();
     auto mapUdfLogicalOperatorNode2 =
         std::make_shared<MapJavaUdfLogicalOperatorNode>(javaUdfDescriptor1, Util::getNextOperatorId());