+++ nes-core/tests/UnitTests/Query/ProjectionTest.cpp
@@ -12,8 +12,6 @@
     limitations under the License.
 */
 
-#include "gtest/gtest.h"
-
 #include <API/QueryAPI.hpp>
 #include <API/Schema.hpp>
 #include <Catalogs/Source/PhysicalSource.hpp>
@@ -350,6 +348,7 @@ TEST_F(ProjectionTest, projectionQueryCorrectField) {
     auto testSourceDescriptor = std::make_shared<TestUtils::TestSourceDescriptor>(
         testSchema,
         [&](OperatorId id,
+            OriginId,
             const SourceDescriptorPtr&,
             const Runtime::NodeEnginePtr&,
             size_t numSourceLocalBuffers,
@@ -417,6 +416,7 @@ TEST_F(ProjectionTest, projectionQueryWrongField) {
     auto testSourceDescriptor = std::make_shared<TestUtils::TestSourceDescriptor>(
         testSchema,
         [&](OperatorId id,
+            OriginId,
             const SourceDescriptorPtr&,
             const Runtime::NodeEnginePtr&,
             size_t numSourceLocalBuffers,
@@ -484,6 +484,7 @@ TEST_F(ProjectionTest, projectionQueryTwoCorrectField) {
     auto testSourceDescriptor = std::make_shared<TestUtils::TestSourceDescriptor>(
         testSchema,
         [&](OperatorId id,
+            OriginId,
             const SourceDescriptorPtr&,
             const Runtime::NodeEnginePtr&,
             size_t numSourceLocalBuffers,
@@ -554,6 +555,7 @@ TEST_F(ProjectionTest, projectOneExistingOneNotExistingField) {
     auto testSourceDescriptor = std::make_shared<TestUtils::TestSourceDescriptor>(
         testSchema,
         [&](OperatorId id,
+            OriginId,
             const SourceDescriptorPtr&,
             const Runtime::NodeEnginePtr&,
             size_t numSourceLocalBuffers,
@@ -601,6 +603,7 @@ TEST_F(ProjectionTest, tumblingWindowQueryTestWithProjection) {
     auto windowSourceDescriptor = std::make_shared<TestUtils::TestSourceDescriptor>(
         windowSchema,
         [&](OperatorId,
+            OriginId,
             const SourceDescriptorPtr&,
             const Runtime::NodeEnginePtr&,
             size_t,
@@ -769,6 +772,7 @@ TEST_F(ProjectionTest, mergeQuery) {
     auto testSourceDescriptor = std::make_shared<TestUtils::TestSourceDescriptor>(
         testSchema,
         [&](OperatorId id,
+            OriginId,
             const SourceDescriptorPtr&,
             const Runtime::NodeEnginePtr&,
             size_t numSourceLocalBuffers,