+++ nes-runtime/src/Execution/Aggregation/MinAggregation.cpp
@@ -44,5 +44,6 @@ void MinAggregationFunction::reset(Nautilus::Value<Nautilus::MemRef> memref) {
     auto minVal = Nautilus::Value<Nautilus::Int64>((int64_t) std::numeric_limits<int64_t>::max());
     memref.store(minVal);// TODO 3280 check the type
 }
+uint64_t MinAggregationFunction::getSize() { return sizeof(int64_t); }
 
 }// namespace NES::Runtime::Execution::Aggregation