+++ nes-runtime/src/Execution/Operators/ExecutionContext.cpp
@@ -95,8 +95,7 @@ Value<MemRef> ExecutionContext::getGlobalOperatorHandler(uint64_t handlerIndex) 
     return FunctionCall<>("getGlobalOperatorHandlerProxy", getGlobalOperatorHandlerProxy, pipelineContext, handlerIndexValue);
 }
 
-Value<MemRef> ExecutionContext::getPipelineContext() const { return pipelineContext; }
-
-Value<MemRef> ExecutionContext::getWorkerContext() const { return workerContext; }
+const Value<MemRef>& ExecutionContext::getWorkerContext() const { return workerContext; }
+const Value<MemRef>& ExecutionContext::getPipelineContext() const { return pipelineContext; }
 
 }// namespace NES::Runtime::Execution